Adicionado a validação de dados nos campos dos formularios, mas do lado do back-end.
[ticket: X]

+++ app/admin/routesFuncionarios.py
@@ -11,30 +11,41 @@ def funcionarios():
 	form = FuncionarioForm()
 	if request.method == 'POST' and form.salvar.data == True:
 		try:
-			user = User(matricula=form.matricula.data, username=form.nome.data, password=form.password.data, is_admin=form.admin.data)
-			db.session.add(user)
-			db.session.commit()
-			flash('Registrado com sucesso', 'success')
+			if form.validacao(form):
+				flash('Falta preenchar um campo!', 'danger')
+			elif form.validInteger(form.matricula.data):
+				flash('Matricula deve ser composto de somente numeros!', 'danger')
+			else:
+				user = User(matricula=form.matricula.data, username=form.nome.data, password=form.password.data, is_admin=form.admin.data)
+				db.session.add(user)
+				db.session.commit()
+				flash('Registrado com sucesso', 'success')
 		except:
 			db.session.rollback()
 			flash('Registro falhou na adição', 'danger')
 			
 	elif request.method == 'GET' and request.args.get('delete'):
 		try:
-			user = User.query.filter_by(matricula=request.args.get('delete')).first_or_404()
-			db.session.delete(user)
-			db.session.commit()
-			flash('Registro apagado com sucesso', 'danger')
+			if request.args.get('delete') == '':
+				flash('Falta a matricula!', 'danger')
+			else:
+				user = User.query.filter_by(matricula=request.args.get('delete')).first_or_404()
+				db.session.delete(user)
+				db.session.commit()
+				flash('Registro apagado com sucesso', 'danger')
 		except:
 			db.session.rollback()
 			flash('Registro falhou em apagar', 'danger')
 
 	elif request.method == 'GET' and request.args.get('update'):
-		user = User.query.filter_by(matricula=request.args.get('update')).first_or_404()
-		form.matricula.data = user.matricula
-		form.nome.data = user.username
-		form.admin.data = user.is_admin
-		return render_template('admin/editFuncionario.html',form=form, action='funcUpdate')
+		if request.args.get('update') == '':
+			flash('Falta a matricula!', 'danger')
+		else:
+			user = User.query.filter_by(matricula=request.args.get('update')).first_or_404()
+			form.matricula.data = user.matricula
+			form.nome.data = user.username
+			form.admin.data = user.is_admin
+			return render_template('admin/editFuncionario.html',form=form, action='funcUpdate')
 
 	listTable=User.query.all()
 	return render_template('admin/funcionarios.html', form=form, listTable=listTable)
@@ -44,16 +55,21 @@ def funcionarios():
 def funcUpdate():
 	form = FuncionarioForm()
 	if request.method == 'POST' and form.salvar.data == True:
-		user = User.query.filter_by(matricula=form.matricula.data).first_or_404()
-		if user:
-			user.username = form.nome.data
-			user.matricula = form.matricula.data
-			user.is_admin = form.admin.data
-			user.password_hash = form.password.data
-			try:
-				db.session.commit()
-				flash('Registro alterado com sucesso', 'info')
-			except:
-				db.session.rollback()
-				flash('Registro falhou em alterar', 'danger')
+		if form.validacao(form):
+			flash('Falta preenchar um campo!', 'danger')
+		elif form.validInteger(form.matricula.data):
+			flash('Matricula deve ser composto de somente numeros!', 'danger')
+		else:
+			user = User.query.filter_by(matricula=form.matricula.data).first_or_404()
+			if user:
+				user.username = form.nome.data
+				user.matricula = form.matricula.data
+				user.is_admin = form.admin.data
+				user.password_hash = form.password.data
+				try:
+					db.session.commit()
+					flash('Registro alterado com sucesso', 'info')
+				except:
+					db.session.rollback()
+					flash('Registro falhou em alterar', 'danger')
 	return redirect(url_for('.funcionarios'))
